Feat/be 24 like endpoint

+++ app/routers/questions.py
@@ -1,9 +1,10 @@
-from fastapi import APIRouter, Depends, status
+from fastapi import APIRouter, Depends, status, HTTPException
 from sqlalchemy.orm import Session
 from .. import model
 from app.database import engine
 from .. import schema
 from app.database import get_db
+from app import oauth
 
 router = APIRouter(
     prefix='/questions',
@@ -12,45 +13,47 @@ router = APIRouter(
     )
 
 
-
 @router.post("/", status_code=status.HTTP_201_CREATED)
-def add_question(request: schema.Question, db: Session = Depends(get_db)):
-    ask_question = model.Question(content=request.content,
-                                  body=request.answered, user_id=request.user_id)
+def add_question(request: schema.Question, db: Session = Depends(get_db), current_user: int = Depends(oauth.get_current_user)):
+    request.owner_id = current_user.user_id
+    ask_question = model.Question(
+        content=request.content, owner_id=request.owner_id)
     db.add(ask_question)
     db.commit()
     db.refresh(ask_question)
     return {"success": True, "message": ask_question.content}
 
 
-@router.patch("/question/{question_id}", status_code=status.HTTP_200_OK)
-def answer_question(question_id, request: schema.Question, db: Session = Depends(get_db)):
-    update_answer = db.query(model.Question).filter(
-        model.Question.id == question_id).first
-    if update_answer:
-        update_answer.answered = request.answered
-        update_answer.user_id = request.user_id
-        update_answer.answer_id = request.answer_id
-        db.commit()
-        db.close()
-        return {"success": True, "message": update_answer.content}
+# @router.patch("/question/{question_id}", status_code=status.HTTP_200_OK)
+# def answer_question(question_id, request: schema.Question, db: Session = Depends(get_db)):
+#     update_answer = db.query(model.Question).filter(
+#         model.Question.question_id == question_id).first
+#     if update_answer:
+#         update_answer.answered = request.answered
+#         update_answer.user_id = request.user_id
+#         update_answer.answer_id = request.answer_id
+#         db.commit()
+#         db.close()
+#         return {"success": True, "message": update_answer.content}
 
 
-@router.delete("/questionid/{question_id}", status_code=status.HTTP_204_NO_CONTENT)
-def delete_question(question_id, db: Session = Depends(get_db)):
+@router.delete("/{question_id}", status_code=status.HTTP_204_NO_CONTENT)
+def delete_question(question_id, db: Session = Depends(get_db), current_user: int = Depends(oauth.get_current_user)):
     delete_question = db.query(model.Question).filter(
         model.Question.id == question_id).first
     if delete_question:
-        db.delete(delete_question)
-        db.commit()
-        return {"success": True, "message": delete_question.question_id}
+        if delete_question.owner_id == current_user.user_id:
+            db.delete(delete_question)
+            db.commit()
+            return {"success": True, "message": delete_question.question_id}
+        else:
+            return {"success": False, "message": "Not authorized"}
 
 
-@router.get("/question/{question_id}", status_code=status.HTTP_200_OK)
-def get_question(question_id, db: Session = Depends(get_db)):
+@router.get("/{question_id}", status_code=status.HTTP_200_OK)
+def get_question(question_id: int, db: Session = Depends(get_db)):
     get_question_db = db.query(model.Question).filter(
-        model.Question.id == question_id).first
-    print(dir(get_question))
+        model.Question.question_id == question_id).first()
     if get_question_db:
         db.commit()
         return {"success": True, "data": {
@@ -64,13 +67,15 @@ def get_question(question_id, db: Session = Depends(get_db)):
         }
 
 
-@router.patch("/question/{question_id}", status_code=status.HTTP_200_OK)
-def update_question(question_id, request: schema.Question, db: Session = Depends(get_db)):
+@router.patch("/{question_id}", status_code=status.HTTP_200_OK)
+def update_question(question_id, request: schema.Question, db: Session = Depends(get_db), current_user: int = Depends(oauth.get_current_user)):
     update_question = db.query(model.Question).filter(
-        model.Question.id == question_id).first
+        model.Question.question_id == question_id).first()
     if update_question:
+        if update_question.owner_id != current_user.user_id:
+            return HTTPException(status_code=404, detail="Not authorized")
         update_question.content = request.content
-        update_question.answer = request.answer
+        update_question.answered = request.answered
         db.commit()
         return {"success": True, "message": update_question.content}
